% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_for_subscripts_cmajor}
\alias{index_for_subscripts_cmajor}
\title{index_for_subscripts_cmajor}
\usage{
index_for_subscripts_cmajor(D, S)
}
\arguments{
\item{D}{Array of dimensions}

\item{S}{Array of subscripts}
}
\value{
linear index of element
}
\description{
Compute the index corresponding to a given
ordered set of subscripts for an array in
first-fastest ("column-major") order.
}
