% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_for_subscripts_rmajor}
\alias{index_for_subscripts_rmajor}
\title{index_for_subscripts_rmajor}
\usage{
index_for_subscripts_rmajor(D, S)
}
\arguments{
\item{D}{Array of dimensions}

\item{S}{Array of subscripts}
}
\value{
linear index of element
}
\description{
Compute the index corresponding to a given
ordered set of subscripts for an array in
last-fastest ("row-major") order.
}
