% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subscripts_for_index_rmajor}
\alias{subscripts_for_index_rmajor}
\title{subscripts_for_index_rmajor}
\usage{
subscripts_for_index_rmajor(D, ind)
}
\arguments{
\item{D}{Array of dimensions}

\item{ind}{Index from beginning of linear array}
}
\value{
Array of subscripts
}
\description{
Compute the subscripts corresponding to a given 
linear index for an array in last-fastest 
("row-major") order.
}
